requirements.txt
# Claude Agent SDK
claude-agent-sdk>=0.1.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Data Validation
pydantic>=2.10.0

# Async Support
asyncio-throttle>=1.0.0

# Code Analysis Tools
radon>=6.0  # Cyclomatic complexity
pylint>=3.0
bandit>=1.7  # Security linter

# Git Operations
gitpython>=3.1

# Testing
pytest>=8.0
pytest-asyncio>=0.23

# Utilities
python-dotenv>=1.0